Wrap signatures onto several lines when function len is over a treshold and function has the focus (#831)
Parameters html are divided into .rst-sig-param or .rst-sig-symbol spans.
When the function is long enough (88 chars excluding the def and ending colon, this is hard coded) an extra CSS class .long-signature is added to the parent function-signature.
The first parameter 'cls' or 'self' of (class) methods is marked with the 'undocumented' CSS class, this way it's clearly not part of the API.
Add some CSS to expand the signature of long functions when they have the focus.
Refactor CSS rules for parameters tables to be more simpler and predictable with percentage values for smaller screens.
Remove ValueFormatter classes since they were only a hack to show HTML tags inside Signature.__str__(), which we don't need anymore because we're computing the representation ourselve.
Remove ParsedStanOnly class because we know have better alternatives.
Introduce ParsedDocstring.to_text() -> str.
When a function has invalid parameters its rendered as (…) instead of confusing () - like it was a zero argument callable.

Co-authored-by: tristanlatr <tristanlatr@users.noreply.github.com>
Co-authored-by: Glyph <glyph@twistedmatrix.com>